+++ src/app.ts
@@ -1,7 +1,7 @@
 import { LitElement, css, html } from "lit";
 import { customElement, state } from "lit/decorators.js";
 import { classMap } from "lit/directives/class-map.js";
-import "./button.js";
+import "./ui/button.js";
 import "./settings/index.js";
 import "./tuner.js";
 import { commonStyles } from "./style.js";
