+++ src/settings/index.ts
@@ -1,6 +1,6 @@
 import { LitElement, css, html } from "lit";
 import { customElement } from "lit/decorators.js";
-import "../button.js";
+import "../ui/button.js";
 import "./temperament.js";
 import "./theme.js";
 import { commonStyles } from "../style.js";
