+++ src/settings/theme.ts
@@ -1,7 +1,7 @@
 import { LitElement, css, html } from "lit";
 import { customElement } from "lit/decorators.js";
 import { classMap } from "lit/directives/class-map.js";
-import "../button.js";
+import "../ui/button.js";
 import { commonStyles } from "../style.js";
 
 export type ThemeName = "system" | "light" | "dark";
